# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "APIs to manage contacts.",
        "title": "Contact API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/contact": {
            "post": {
                "description": "This endpoint is used to create a new contact",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contact"
                ],
                "summary": "Creates new Contact",
                "parameters": [
                    {
                        "description": "contact info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_api_dto.ContactRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_api_dto.ContactResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_shared_errorhandler.TemplateError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_shared_errorhandler.TemplateError"
                        }
                    }
                }
            }
        },
        "/contact/{contactId}": {
            "get": {
                "description": "Get contact by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contact"
                ],
                "summary": "Get contact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "contact ID",
                        "name": "contactId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_api_dto.ContactResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_shared_errorhandler.TemplateError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_shared_errorhandler.TemplateError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_shared_errorhandler.TemplateError"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the Contact",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contact"
                ],
                "summary": "Update Contacts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contact ID",
                        "name": "contactId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Contact request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_api_dto.ContactRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_api_dto.ContactResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_shared_errorhandler.TemplateError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_shared_errorhandler.TemplateError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_shared_errorhandler.TemplateError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Contact by ID",
                "tags": [
                    "Contact"
                ],
                "summary": "Delete Contact by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "contact ID",
                        "name": "contactId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_shared_errorhandler.TemplateError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_shared_errorhandler.TemplateError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_josue_challenge-accel-one_internal_shared_errorhandler.TemplateError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_josue_challenge-accel-one_internal_api_dto.ContactRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                }
            }
        },
        "github_com_josue_challenge-accel-one_internal_api_dto.ContactResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                }
            }
        },
        "github_com_josue_challenge-accel-one_internal_shared_errorhandler.TemplateError": {
            "type": "object",
            "properties": {
                "error": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}